[CI] #61 Linuxデスクトップ非対応のfirebase_authをCIから分離
firebase_auth / cloud_firestore がLinuxデスクトップでコンパイル不可なため、
AuthServiceインターフェースを抽出しFirebase実装をProviderScope経由で注入する
構成に変更。CIはFirebase非依存のテストパスのみ明示的に実行する。

- auth_provider.dart からFirebaseを除去、AuthUser/AuthException を定義
- FirebaseAuthServiceImpl を firebase_auth_service_impl.dart に分離
- notificationServiceProvider を notification_provider.dart に分離
- main_*.dart でProviderScope.overrides により実装を注入
- CI yml: Firebase非依存テストパスのみ列挙、integration_test除外
